requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
datasets==2.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docutils==0.20.1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.109.2
fastjsonschema==2.18.1
ffmpy==0.3.2
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2023.9.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
gradio==4.18.0
gradio_client==0.10.0
greenlet==2.0.2
grpcio==1.59.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
imageio==2.33.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyterlab-pygments==0.2.2
keyring==24.2.0
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nh3==0.2.14
numpy==1.26.0
oauthlib==3.2.2
openai==1.10.0
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
orjson==3.9.13
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pkginfo==1.9.6
platformdirs==3.11.0
playwright==1.38.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.6.0
pydantic_core==2.16.1
pydub==0.25.1
pyee==9.0.4
Pygments==2.16.1
pyloco==0.0.139
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
readme-renderer==42.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
ruff==0.2.1
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
semantic-version==2.10.0
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.16.0
sklearn==0.0.post9
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchinfo==1.8.0
torchmetrics==1.2.1
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
twine==4.0.2
typer==0.9.0
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
ushlex==0.99.1
uvicorn==0.27.1
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
